chore(worker): add graph module to workflow-runner crate
