+++ worker/crates/action-universal/Cargo.toml
@@ -17,7 +17,6 @@ reearth-flow-common.workspace = true
 reearth-flow-eval-expr.workspace = true
 reearth-flow-geometry.workspace = true
 reearth-flow-storage.workspace = true
-reearth-flow-workflow.workspace = true
 reearth-flow-xml.workspace = true
 
 nusamai-citygml.workspace = true
