+++ worker/crates/action/Cargo.toml
@@ -16,7 +16,6 @@ reearth-flow-common.workspace = true
 reearth-flow-eval-expr.workspace = true
 reearth-flow-geometry.workspace = true
 reearth-flow-storage.workspace = true
-reearth-flow-workflow.workspace = true
 
 nusamai-citygml.workspace = true
 nusamai-geometry.workspace = true
