+++ worker/crates/cli/Cargo.toml
@@ -21,7 +21,6 @@ reearth-flow-state.workspace = true
 reearth-flow-storage.workspace = true
 reearth-flow-telemetry.workspace = true
 reearth-flow-workflow-runner.workspace = true
-reearth-flow-workflow.workspace = true
 
 bytes.workspace = true
 clap = {version = "4.5.4", features = ["env", "string"]}
