+++ worker/crates/cli/src/run.rs
@@ -8,8 +8,8 @@ use reearth_flow_action_log::factory::{create_root_logger, LoggerFactory};
 use reearth_flow_common::uri::Uri;
 use reearth_flow_state::State;
 use reearth_flow_storage::resolve;
-use reearth_flow_workflow::{id::Id, workflow::Workflow};
 use reearth_flow_workflow_runner::dag::DagExecutor;
+use reearth_flow_workflow_runner::types::{graph::Id, graph::Workflow};
 
 pub fn build_run_command() -> Command {
     Command::new("run")
