+++ worker/crates/workflow-runner/Cargo.toml
@@ -18,9 +18,9 @@ reearth-flow-common.workspace = true
 reearth-flow-eval-expr.workspace = true
 reearth-flow-state.workspace = true
 reearth-flow-storage.workspace = true
-reearth-flow-workflow.workspace = true
 
 async-recursion.workspace = true
+bincode.workspace = true
 once_cell.workspace = true
 opentelemetry.workspace = true
 petgraph.workspace = true
