+++ worker/crates/workflow-runner/src/lib.rs
@@ -1,5 +1,6 @@
 mod action_runner;
 pub mod dag;
+pub mod types;
 
 pub type Result<T, E = Error> = std::result::Result<T, E>;
