+++ worker/crates/workflow-runner/src/types/graph.rs
@@ -1,10 +1,36 @@
 use serde::{Deserialize, Serialize};
 use serde_json::{Map, Value};
+use uuid::Uuid;
 
-use crate::id::Id;
+use reearth_flow_common::serde::from_str;
 
+pub type Id = Uuid;
 pub type NodeProperty = Map<String, Value>;
 pub type NodeAction = String;
+pub type Parameter = Map<String, Value>;
+
+#[derive(Serialize, Deserialize, Debug)]
+#[serde(rename_all = "camelCase")]
+pub struct Workflow {
+    pub id: Id,
+    pub name: String,
+    pub entry_graph_id: Id,
+    pub with: Option<Parameter>,
+    pub graphs: Vec<Graph>,
+}
+
+#[derive(Serialize, Deserialize, Debug)]
+#[serde(rename_all = "camelCase")]
+pub struct WorkflowParameter {
+    pub global: Option<Parameter>,
+    pub node: Option<NodeProperty>,
+}
+
+impl Workflow {
+    pub fn try_from_str(s: &str) -> crate::Result<Self> {
+        from_str(s).map_err(|e| crate::Error::Init(format!("Failed to parse workflow: {}", e)))
+    }
+}
 
 #[derive(Serialize, Deserialize, Debug, Clone)]
 pub struct NodeEntity {
