--- worker/crates/workflow/Cargo.toml
@@ -1,21 +1,0 @@
-[package]
-description = "Re:Earth Flow workflow"
-name = "reearth-flow-workflow"
-
-authors.workspace = true
-edition.workspace = true
-homepage.workspace = true
-license.workspace = true
-repository.workspace = true
-rust-version.workspace = true
-version.workspace = true
-
-[dependencies]
-reearth-flow-common.workspace = true
-
-serde.workspace = true
-serde_json.workspace = true
-tracing-subscriber.workspace = true
-tracing.workspace = true
-thiserror.workspace = true
-uuid.workspace = true
