--- worker/crates/workflow/src/id.rs
@@ -1,3 +1,0 @@
-use uuid::Uuid;
-
-pub type Id = Uuid;
