+++ worker/examples/Cargo.toml
@@ -21,7 +21,6 @@ reearth-flow-eval-expr.workspace = true
 reearth-flow-state.workspace = true
 reearth-flow-storage.workspace = true
 reearth-flow-workflow-runner.workspace = true
-reearth-flow-workflow.workspace = true
 
 bytes.workspace = true
 directories.workspace = true
