+++ worker/tests/helper.rs
@@ -6,8 +6,8 @@ use reearth_flow_action_log::factory::LoggerFactory;
 use reearth_flow_common::uri::Uri;
 use reearth_flow_state::State;
 use reearth_flow_storage::resolve::StorageResolver;
-use reearth_flow_workflow::workflow::Workflow;
 use reearth_flow_workflow_runner::dag::DagExecutor;
+use reearth_flow_workflow_runner::types::graph::Workflow;
 
 #[derive(RustEmbed)]
 #[folder = "fixture/testdata/"]
